Move everything into the outer layout to have it also for eg news pages
